chore: Update view paths in config file and blade templates
This commit updates the view paths in the `config/view.php` file and the blade templates in the `resources/views/backend` directory. The view paths were changed from using the `layouts` namespace to the `backend.layouts` namespace. This change ensures that the correct master layout files are used for the backend views.